feat: improve button component
