+++ Main/Scenes/Common/Theme/Colors.swift
@@ -3,8 +3,8 @@ import UIKit.UIColor
 extension UIColor {
   struct Cora {
     static let primaryColor = UIColor(named: "AccentColor")!
-    static let onPrimaryColor = UIColor(named: "OnAccentColor")
-    static let primaryBgColor = UIColor(named: "SecondaryBgColor")
+    static let onPrimaryColor = UIColor(named: "OnAccentColor")!
+    static let primaryBgColor = UIColor(named: "SecondaryBgColor")!
     static let secondaryBgColor = UIColor(named: "SecondaryBgColor")!
     static let highlightedTextColor = UIColor(named: "HighlightedTextColor")!
   }
