+++ Main/Scenes/Common/UIColor+Darker.swift
@@ -1,13 +1,2 @@
 import UIKit.UIColor
 
-extension UIColor {
-    func darker() -> UIColor {
-        var r:CGFloat = 0, g:CGFloat = 0, b:CGFloat = 0, a:CGFloat = 0
-
-        if self.getRed(&r, green: &g, blue: &b, alpha: &a){
-            return UIColor(red: max(r - 0.2, 0.0), green: max(g - 0.2, 0.0), blue: max(b - 0.2, 0.0), alpha: a)
-        }
-
-        return UIColor()
-    }
-}
