--- Main/Scenes/Common/UIView+Constraints.swift
@@ -1,8 +1,0 @@
-import UIKit
-
-extension UIView {
-  func makeConstraints(_ completion: (UIView) -> [NSLayoutConstraint]) {
-    self.translatesAutoresizingMaskIntoConstraints = false
-    NSLayoutConstraint.activate(completion(self))
-  }
-}
